requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# VLM API SDKs
anthropic>=0.40.0
openai>=1.50.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# HTTP client with async support
httpx>=0.25.0

# Audio processing
pyaudio>=0.2.13
pyttsx3>=2.90  # TTS (offline)
SpeechRecognition>=3.10.0  # STT
webrtcvad>=2.0.10  # Voice activity detection

# Utilities
pydantic>=2.0.0
tenacity>=8.2.0

# Existing project dependencies (preserve)
PyQt5>=5.15.0
picamera2>=0.3.0
gpiozero>=2.0.0
smbus>=1.1.0
rpi-ws281x>=5.0.0
